pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "graph-link-prediction"
version = "0.1.0"
description = "Hybrid graph link prediction with structural and semantic embeddings"
readme = "README.md"
authors = [
  { name = "Radoslaw Kowalczyk", email = "r.kowalczyk@outlook.com" }
]
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "networkx>=3.1",
  "matplotlib>=3.7",
  "scikit-learn>=1.3",
  "torch>=2.2",
  "transformers>=4.44",
  "torch-geometric>=2.6",
]

[project.optional-dependencies]
test = [
  "pytest>=7",
  "pytest-cov>=4.1",
]
dev = [
  "pre-commit>=3.7",
]

[project.urls]
Homepage = "https://github.com/yourname/graph-link-prediction"
Issues = "https://github.com/yourname/graph-link-prediction/issues"

[tool.hatch.build.targets.wheel]
packages = ["graph_lp"]

[tool.ruff]
extend-exclude = ["notebooks/**"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
follow_imports = "skip"
warn_unused_ignores = false
allow_redefinition = true

[[tool.mypy.overrides]]
module = ["graph_lp.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["notebooks.*"]
ignore_errors = true
